🎨 Remove unused dependency from origin-web-console
The recommendation from @spadgett is to stick to services in
origin-web-common as they are less likely to change.
Using services from origin-web-console would have a higher chance
of breaking UI extensions in newer versions of openshift.

This change removes an unsued service that is *not* part of
origin-web-common. There is no risk from this change