✨List Provisioned Services in the Mobile Overview
This change removes the 'Mobile Services' & 'Services' areas on the
overview screen in favour of a single 'Provisioned Services' area.
All services (including Mobile specific) will be shown here.

A potential followup could be to split the mobile & non-mobile services
again, but this change should be sufficient for the POC